// src/enums/enums.ts
export enum SizesEnum {
  S = 's',
  M = 'm',
  L = 'l',
}
// ...
export enum FontsEnum {
  Text12 = '400-12',
  Text14 = '400-14',
  Text16 = '400-16',
  Text18 = '400-18',

  Label14 = '500-14',
  Label16 = '500-16',
  Label18 = '500-18',
  Label20 = '500-20',
  Label30 = '500-30',
  Label40 = '500-40',
  Label50 = '500-50',

  Header14 = '700-14',
  Header16 = '700-16',
  Header18 = '700-18',
  Header20 = '700-20',
  Header22 = '700-22',
  Header26 = '700-26',
  Header30 = '700-30',
}